Added auth

+++ backend/src/reservations/reservations.controller.ts
@@ -9,19 +9,25 @@ import {
   UsePipes,
   ValidationPipe,
   Req,
+  UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
 import { ReservationsService } from './reservations.service';
 import { CreateReservationDto } from './dto/create-reservation.dto';
+import { ClientGuard } from '../auth/guards/client.guard';
+import { ManagerGuard } from '../auth/guards/manager.guard';
+import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
 
 @Controller()
 export class ReservationsController {
   constructor(private readonly reservationsService: ReservationsService) {}
 
   @Post('/api/client/reservations')
+  @UseGuards(JwtAuthGuard, ClientGuard)
   @UsePipes(new ValidationPipe({ whitelist: true }))
   async create(@Body() dto: CreateReservationDto, @Req() req: any) {
-    // const userId = req.user?.id || 'temp-user-id';
-    const userId = '691b18ea2dd236656e2dd459';
+    const userId = req.user.userId;
+
     const room = await this.reservationsService.getRoomById(dto.hotelRoom);
     const hotelId = (room.hotel as any)._id.toString();
 
@@ -53,8 +59,9 @@ export class ReservationsController {
   }
 
   @Get('/api/client/reservations')
+  @UseGuards(JwtAuthGuard, ClientGuard)
   async getUserReservations(@Req() req: any) {
-    const userId = req.user?.id || 'temp-user-id';
+    const userId = req.user.userId;
 
     const reservations = await this.reservationsService.getReservations({
       userId,
@@ -75,12 +82,21 @@ export class ReservationsController {
   }
 
   @Delete('/api/client/reservations/:id')
+  @UseGuards(JwtAuthGuard, ClientGuard)
   async cancelReservation(@Param('id') id: string, @Req() req: any) {
+    const reservation =
+      await this.reservationsService.getReservationWithDetails(id);
+
+    if (reservation.userId.toString() !== req.user.userId) {
+      throw new ForbiddenException('You can only cancel your own reservations');
+    }
+
     await this.reservationsService.removeReservation(id);
     return {};
   }
 
   @Get('/api/manager/reservations/:userId')
+  @UseGuards(JwtAuthGuard, ManagerGuard)
   async getUserReservationsManager(@Param('userId') userId: string) {
     const reservations = await this.reservationsService.getReservations({
       userId,
@@ -101,6 +117,7 @@ export class ReservationsController {
   }
 
   @Delete('/api/manager/reservations/:id')
+  @UseGuards(JwtAuthGuard, ManagerGuard)
   async cancelReservationManager(@Param('id') id: string) {
     await this.reservationsService.removeReservation(id);
     return {};
